Add simple almost-universal multiply shift.
The multiply-shift already prsent was (H1*k + H2) >> w, where k was w
bits wide and H1 and H2 were 2k bits wide.

This is a simpler hash family H1*k, where H1 is the same width as k
and H1 must be odd. It is not 2-independent, but it is
2-almost-universal and it uses 1/4 the randomness of the other
multiply-shift scheme, as well as fewer operations.