Add a "split" hasher.
This hash family is built from two other hash families, a Splitter and
a Finalizer. It keeps N different hash functions from the Finalizer
family. It uses the splitter to hash the key to a value v in [0,N),
then uses the v'th finalizer to hash the key.

It performs well in the bucket benchmark, even when instatiated with
families for Splitter and Finalizer that do not do so well.